Add CertificateStore interface
CertificateStore enables different storage backends for certificates. The methods in the interface were based on CertificateStore (now called SQLiteCertificateStore). Method names were shortened since the context in which they are called makes their verbose naming redundant. With regard to dependency injection, a DataBindings interface was created to inject implementations of interfaces. In addition, the constructor of SQLiteCertificateStore was annotate with @Inject to enable DI binding.

Signed-off-by: rtatton <rdt17@protonmail.com>